[LW-10570] Fix hardware wallets tx logic
